+++ src/pixels.rs
@@ -33,22 +33,12 @@ where
             let color = match self.raw.raw_bmp.color_bpp() {
                 // 1 and 8 BPP images may use a color table if one is provided
                 Bpp::Bits1 | Bpp::Bits8 => {
-                    self.raw
-                        .raw_bmp
-                        .color_table()
-                        .map(|table| {
-                            // Each color table entry is 4 bytes long
-                            let offset = color as usize * 4;
+                    let table = self.raw.raw_bmp.color_table()?;
 
-                            u32::from_le_bytes([
-                                table[offset + 0],
-                                table[offset + 1],
-                                table[offset + 2],
-                                table[offset + 3],
-                            ])
-                        })
-                        // No color mapping - use pixel value directly
-                        .unwrap_or(color)
+                    // Each color table entry is 4 bytes long
+                    let offset = color as usize * 4;
+                    
+                    u32::from_le_bytes(table[offset..offset + 4].try_into().unwrap())
                 }
                 // Color table should be ignored for any other bit depth
                 _ => color,